refact: pricing hook 분리
