+++ src/advanced/__tests__/advanced.test.tsx
@@ -1,4 +1,4 @@
-import {describe, expect, it} from "vitest";
+import { describe, expect, it } from 'vitest';
 
 describe('advanced test', () => {
   it('테스트코드를 자유롭게 한번 연습해보세요.', () => {
